Add Schema#headers
Also add Data#schema, Data#fields, and Data#headers.
