+++ google-cloud-bigquery/lib/google/cloud/bigquery/view.rb
@@ -300,7 +300,7 @@ module Google
         # @!group Attributes
         #
         def headers
-          fields.map(&:name).map(&:to_sym)
+          schema.headers
         end
 
         ##
